Created authorization with gates in AppServiceProvider, added auth to Review and Book Controllers

+++ app/Http/Controllers/Api/BookController.php
@@ -6,9 +6,11 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\BookResource;
 use App\Models\Book;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
+
 
     public function __construct() {
         $this->middleware("auth:sanctum")->except(['index', 'show']);
@@ -24,6 +26,7 @@ class BookController extends Controller
     public function store(Request $request)
     {
         $book = Book::create([
+            "user_id" => $request->user()->id,
             ...$request->validate([
                 'title' => 'required|string|max:50',
                 'author' => 'required|string|max:50',
@@ -42,6 +45,10 @@ class BookController extends Controller
 
     public function update(Request $request, Book $book)
     {
+
+
+        $this->authorize('update-book', $book);
+
         $book->update($request->validate([
             'title' => 'sometimes|string|max:50',
             'author' => 'sometimes|string|max:50',
@@ -54,6 +61,8 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+
+        $this->authorize('delete-book', $book);
 
         $book->delete();
         return response(status: 204);
